0.5.11 updates in preparation for 0.6.0

+++ acinclude.m4
@@ -28,11 +28,12 @@ AC_DEFUN(AC_CPP_PIPE_STDIN,
 	fi])
 
 AC_DEFUN(AC_UPDATE_IF_CHANGED,
-	[if cmp -s "$1" $2 2>/dev/null; then
-		echo "$1 is unchanged"
-		rm -f "$2"
-	else
-		echo "$1 has changed"
-		rm -f "$1"
-		mv "$2" "$1"
+	[if test -f "$2"; then
+		if cmp -s "$1" "$2" 2>/dev/null; then
+			echo "$1 is unchanged"
+		else
+			echo "$1 has changed"
+			rm -f "$1"
+			mv "$2" "$1"
+		fi
 	fi])
